remove mnemonic and "episode assistant" module
In the beginning, everything evolved around the episode numbers and the
mnemonic. Then, it made sense to support this by something like the
episode assistant.

Now, the mnemonic is merely an afterthought. It's used by no part of
the system except the episode assistant. And this doesn't do a lot that
can't be done without it either. So we decided to drop both for now.

A similar concept might return once we tackle stuff like seasons and so
on.