enforce post_status=publish in permalink proxy
why? when migrating, there are 2 posts with the same slug.
the WP query uses the first one it finds. sometimes it's the
correct one (the published episode), sometimes it's the
trashed post. Why would you ever want to fetch something
that is not public when parsing the url?
So ... I just enforce it here, hoping not to break anything.