:bug: fix ADN announcement encoding issue
get_the_title() escapes characters like dashes into
html entities. That messes with the determination of
position and length of linked episodes.

Besides, using get_the_title() in the save hook was
not the best in the first place as it does not use
the current title but the one from the last save
(which will be identical in most cases but it does
not hurt to fix two bugs at once)