fix missing default contributors
`$c->id` was empty for all default contributors, which
resulted in all but the last one being overwritten.
The solution is to use the contributor id instead but
prepend a static string (“default”) to avoid it being misinterpreted as
a contribution id later, which
might result in deletion of totally unrelated contributions.
Anyway, this is also not good enough if one contributor
is used more than once as a default — unlikely but possible.
Hence the introduction of the $i counter to make each
default contribution key truly unique.