slugify: use rawurlencode instead of urlencode
Rule of thumb is: use `rawurlencode` for everything _before_ the `?`
and `urlencode` for everything after. Since `slugify` is used for the
former, this change makes sense. However, the recommended usage is not
obvious from the method name, so … it would be great to find a more
clear solution at some point.

One notable difference is how spaces are handled:

```
$> rawurlencode(" ");
// '%20'
$> urlencode(" ");
// '+'
```