give user agent refresh more time
It turns out that WordPress does not start a new process for each cron
event. So even if I schedule a cron for each 500 UAs, they might all be
executed in the same request and run into timeout.

So I decided to make 500UA-chunks instead and increase the distribution
time, increasing the chances of random clicks by the admin during that
time. It’s not absolutely failsafe, but let’s see if it turns out to be
good enough.