fix duplicate episodes in feeds
For some reason not absolutely clear to me, the post-episode-search
logic has an effect on feed queries (despite there being no search
argument). That results in a JOIN to the episode table within the posts
query that is not required in feeds.

I now specifically exclude feeds from that search logic, which fixes
the bug.