make image resizing async
Resizing is, depending on the machine, quite fast (25-100ms) and I
assumed it would be acceptable to do it on the fly. However, it makes
archive pages basically inaccessible on a “cold” machine. Imagine 100
episodes with 50 different contributors and 25 different social icons.
That’s 175 images, times 3 due to retina, so over 500 on-the-fly
resizes. Not acceptable.

Moving it into WP Cron is much better, but I wish we had a proper
queued background worker system for heavy lifting like that. Maybe it’s
worth looking for drop-in replacements for the cron system or
optionally supporting something like “php resque” in the Publisher.

Tasks that could move into that system:

- Image Cache downloads / resizing
- Tracking / Analytics calculations
- Episode / Asset / Feed validations
- later: conflict system checks