handle valid files with missing Content-Length header
Files with missing Content-Length header are now treated as valid. The
size is displayed as “unknown” and a warning is generated in the log.

Internally, I set the file size to “1 byte” to signal “unknown” size,
since the codebase generally checks for “file size > 0” for validity.
This must be replaced by a sophisticated system. Possibly once we
introduce conflicts … ? One must be allowed to dream.
:rainbow::rainbow::rainbow: