fix current_time_group
In the downloads list table, the current progress jumps to the next
column before the current column is calculated on a purely time-based
approach. This results in a column with a missing number for up to an
hour (depending on when the job runs).

This new approach takes into account what numbers have been calculated
and which are still missing.