fix template bug (when a template returned an empty result, the template title was displayed instead)
I did not anticipate a template returning nothing, but that’s possible
for example if you wrap the whole template in “if not is_feed()” and
then render it in the feed. Before, this counted as a failed render and
the fallback render would take place: assuming the template title is
the HTML. Now an empty result counts as a success.

fixes #932