feat: new "Slacknotes" module
A module that connects to a Slack channel. From that channel, links
are extracted and their metadata (website title and source).

The goal of the module is to generate an HTML list of those links to
be used as show notes.

This final HTML can be copied with a simple click.

To control which links are part of the result, a date range can be
specified and single links can be deselected.

For convenience, the previously chosen Slack channel is remembered.