Update share.html
Hello,

I think I just found a bug in the share.html. It seems like there are some 404 Errors for some scripts and styles if an iFrame is embedded on other pages. So I just added the src path and now it works for me.
I don't know if this a bug or not but after I made those changes the embedded iFrame worked again.

LG