Check if podlovePlayer is defined (#1060)
If the function is undefined (for example in IE11), it may cause problems with other scripts.  
For example, on our site the [a3 Lazy Load plugin](https://wordpress.org/plugins/a3-lazy-load/) ceased working if podlove was embedded on a page.  
Sadly around 25% of our users use Internet Explorer 11, so this affected a large part of our user base.