feat: save/update production
I only wanted to update the tracks, but that does not work because when sending tracks, they are _added_ instead of replaced. There is an endpoint for deleting all tracks, but that also deleted intro/outro files.

So the only option left is using the reset_data route, meaning I need to always send all production data (except uploads), which is harder than one might think, because the production json one gets from the API is not exactly the format required for updatin a production. It's enriched with convenience fields and some null fields (like chapter images/urls) must not be sent as payload.

Not quite there yet but proof of concept is working.